Move GitHub GraphQL methods to utils.py
