+++ src/fetchcode/package_versions.py
@@ -16,7 +16,6 @@
 
 import dataclasses
 import logging
-import os
 import traceback
 import xml.etree.ElementTree as ET
 from datetime import datetime
@@ -30,6 +29,8 @@ from dateutil import parser as dateparser
 from packageurl import PackageURL
 from packageurl.contrib.route import NoRouteAvailable
 from packageurl.contrib.route import Router
+
+from fetchcode.utils import fetch_github_tags_gql
 
 logger = logging.getLogger(__name__)
 
@@ -287,7 +288,8 @@ def get_github_versions_from_purl(purl):
     """Fetch versions of ``github`` packages using GitHub REST API."""
     purl = PackageURL.from_string(purl)
 
-    yield from fetch_github_tags_gql(purl)
+    for version, date in fetch_github_tags_gql(purl):
+        yield PackageVersion(value=version, release_date=date)
 
 
 @router.route("pkg:golang/.*")
@@ -549,124 +551,3 @@ def remove_debian_default_epoch(version):
     ''
     """
     return version and version.replace("0:", "")
-
-
-def fetch_github_tags_gql(purl):
-    """
-    Yield PackageVersion for given github ``purl`` using the GitHub GQL API.
-    """
-    for node in fetch_github_tag_nodes(purl):
-        name = node["name"]
-        target = node["target"]
-
-        # in case the tag is a signed tag, then the commit info is in target['target']
-        if "committedDate" not in target:
-            target = target["target"]
-
-        committed_date = target.get("committedDate")
-        release_date = None
-        if committed_date:
-            release_date = dateparser.parse(committed_date)
-
-        yield PackageVersion(value=name, release_date=release_date)
-
-
-GQL_QUERY = """
-query getTags($name: String!, $owner: String!, $after: String)
-{
-    repository(name: $name, owner: $owner) {
-        refs(refPrefix: "refs/tags/", first: 100, after: $after) {
-            totalCount
-            pageInfo {
-                endCursor
-                hasNextPage
-            }
-            nodes {
-                name
-                target {
-                    ... on Commit {
-                        committedDate
-                    }
-                    ... on Tag {
-                            target {
-                            ... on Commit {
-                                committedDate
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-}"""
-
-
-def fetch_github_tag_nodes(purl):
-    """
-    Yield node name/target mappings for Git tags of the ``purl``.
-
-    Each node has this shape:
-        {
-        "name": "v2.6.24-rc5",
-        "target": {
-            "target": {
-            "committedDate": "2007-12-11T03:48:43Z"
-            }
-        }
-        },
-    """
-    variables = {
-        "owner": purl.namespace,
-        "name": purl.name,
-    }
-    graphql_query = {
-        "query": GQL_QUERY,
-        "variables": variables,
-    }
-
-    while True:
-        response = github_response(graphql_query)
-        refs = response["data"]["repository"]["refs"]
-        for node in refs["nodes"]:
-            yield node
-
-        page_info = refs["pageInfo"]
-        if not page_info["hasNextPage"]:
-            break
-
-        # to fetch next page, we just set the after variable to endCursor
-        variables["after"] = page_info["endCursor"]
-
-
-class GitHubTokenError(Exception):
-    pass
-
-
-class GraphQLError(Exception):
-    pass
-
-
-def github_response(graphql_query):
-    gh_token = os.environ.get("GH_TOKEN", None)
-
-    if not gh_token:
-        msg = (
-            "GitHub API Token Not Set\n"
-            "Set your GitHub token in the GH_TOKEN environment variable."
-        )
-        raise GitHubTokenError(msg)
-
-    headers = {"Authorization": f"bearer {gh_token}"}
-
-    endpoint = "https://api.github.com/graphql"
-    response = requests.post(endpoint, headers=headers, json=graphql_query).json()
-
-    message = response.get("message")
-    if message and message == "Bad credentials":
-        raise GitHubTokenError(f"Invalid GitHub token: {message}")
-
-    errors = response.get("errors")
-    if errors:
-        raise GraphQLError(errors)
-
-    return response
